requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1
google-generativeai==0.8.3
mysql-connector-python==9.1.0
python-multipart==0.0.12
